requirements.txt
google-generativeai>=0.3.2
opencv-python>=4.8.0
moviepy>=1.0.3
pydub>=0.25.1
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
json5>=0.9.14
Pillow>=9.0.0
pygame>=2.0.0
librosa>=0.10.0